requirements.txt
streamlit
pandas
plotly
yfinance
google-genai
python-dotenv
